Implement schedule methods separately.
