+++ public/js/index.js
@@ -15,11 +15,6 @@ let headers = {
 };
 
 
-/* * * * *     Event Triggers     * * * * */
-// form submit for form I-1, student perspective.
-$('#btn-form-i-1-student').on('click', function () {
-    getFormI1StudentDetails();
-});
 // form submit for form I-1, supervisor perspective.
 $('#btn-form-i-1-supervisor').on('click', function () {
     getFormI1SupervisorDetails();
@@ -56,81 +51,6 @@ $('#btn-logout').on('click', function (e) {
 
 
 /* * * * * * * Forms * * * * * * */
-
-/**** Liyanage A.Y.K. *****/
-/*
- * this will get the information filled by the student,
- * on Form I-1 and validate and prepare in order to be sent,
- * to the backend.
- */
-function getFormI1StudentDetails() {
-    let data = {
-        name: document.getElementById('name-student').value,
-        studentId: document.getElementById('id-student').value,
-        address: document.getElementById('address-student').value,
-        homePhone: document.getElementById('home-phone-student').value,
-        mobilePhone: document.getElementById('mobile-phone-student').value,
-        emailAddresses: document.getElementById('emails-student').value, // may contain multiple values separated by comma ( , )
-        year: document.getElementById('year-student').value,
-        semester: document.getElementById('semester-student').value,
-        cgpa: document.getElementById('cgpa-student').value,
-        assignedSupervisor: document.getElementById('assigned-supervisor').value
-    }
-
-
-    /* formatting input parameters. */
-    // getting the multiple emails in an array.
-    data.emailAddresses = data.emailAddresses.includes(',') ? data.emailAddresses.replace(' ', '').split(',') : data.emailAddresses;
-
-    // replacing spaces of IT number since some students may type IT 16 1111 11 instead of IT16111111
-    // splitting by spaces and joining without a space will replace all the spaces since .replace() function only replace one occurrence
-    data.studentId = data.studentId.includes(' ') ? data.studentId.split(' ').join('') : data.studentId;
-
-
-    axios.post(baseUrl + '/forms/form-i-1/student/' + data.studentId, data)
-        .then(response => {
-            console.log(response.data);
-            if (response.data.success) {
-                alert('Data enter to Form I-1 successfully.');
-            }
-        })
-        .catch(error => {
-            console.log(error);
-            alert('Unsuccessful! Please check your data.');
-        })
-}
-
-
-/*
- * this will get the information filled by the supervisor,
- * on Form I-1 and validate and prepare in order to be sent,
- * to the backend.
- */
-function getFormI1SupervisorDetails() {
-    let data = {
-        studentId: document.getElementById('id-student').value, // because supervisor data is stored within the student's record.
-        employerName: document.getElementById('name-employer').value,
-        employerAddress: document.getElementById('address-employer').value,
-        supervisorName: document.getElementById('name-supervisor').value,
-        supervisorTitle: document.getElementById('title-supervisor').value,
-        supervisorPhone: document.getElementById('phone-supervisor').value,
-        supervisorEmail: document.getElementById('email-supervisor').value,
-        internshipStart: document.getElementById('internship-start-date').value,
-        internshipEnd: document.getElementById('internship-end-date').value,
-        workHoursPerWeek: document.getElementById('no-of-hours').value
-    };
-
-    axios.post(baseUrl + '/forms/form-i-1/supervisor/' + data.studentId, data)
-        .then(response => {
-            console.log(response.data);
-            alert(response.data.message)
-        })
-        .catch(error => {
-            console.log(error);
-        })
-}
-
-
 /**** Tharindu TCJ *****/
 
 function populateFormI1() {
